Move reservation controls to shared folder
