+++ app/pages/user-reservations/reservation-list/ReservationListItem.js
@@ -3,7 +3,7 @@ import React, { Component, PropTypes } from 'react';
 import { Link } from 'react-router';
 
 import TimeRange from 'components/common/TimeRange';
-import ReservationControls from 'containers/ReservationControls';
+import ReservationControls from 'shared/reservation-controls';
 import { getCaption, getMainImage } from 'utils/imageUtils';
 import { renderReservationStateLabel } from 'utils/renderUtils';
 import { getResourcePageUrl } from 'utils/resourceUtils';
