+++ app/pages/user-reservations/reservation-list/ReservationListItem.spec.js
@@ -4,11 +4,11 @@ import React from 'react';
 import Immutable from 'seamless-immutable';
 
 import TimeRange from 'components/common/TimeRange';
-import ReservationControls from 'containers/ReservationControls';
 import Image from 'fixtures/Image';
 import Reservation from 'fixtures/Reservation';
 import Resource from 'fixtures/Resource';
 import Unit from 'fixtures/Unit';
+import ReservationControls from 'shared/reservation-controls';
 import { getResourcePageUrl } from 'utils/resourceUtils';
 import ReservationListItem from './ReservationListItem';
