+++ app/shared/reservation-controls/index.js
@@ -1,0 +1,3 @@
+import ReservationControlsContainer from './ReservationControlsContainer';
+
+export default ReservationControlsContainer;
